update rank after get match result

+++ Backend/main/Matching MS/Matching-MS/src/main/java/csd/backend/Matching/MS/MatchmakingQueueListener.java
@@ -171,7 +171,7 @@ public class MatchmakingQueueListener {
             JsonNode rootNode = objectMapper.readTree(messageBody);
 
             Long playerId = rootNode.path("playerId").asLong();
-            int rankId = rootNode.path("rank").asInt();
+            Long rankId = rootNode.path("rank").asLong();
 
             playerService.updatePlayerRank(playerId, rankId);
             System.out.println("Player profile updated for playerId: " + playerId);
